+++ R/rescale.R
@@ -51,15 +51,15 @@ rescale <- function(x, fixed2, scale_pars, MCMC, refs, X2_names, trafos) {
     0
   }
 
-  square <- if (x %in% trafos$var & any(trafos$fct == paste0(x, "^2"))) {
-    qdr_var <- trafos$Xc_var[which(trafos$var == x &
-                                           trafos$fct == paste0(x, "^2"))]
-    2 * as.numeric(MCMC[ , qdr_var])/scale_pars["scale", x]^2 * scale_pars["center", x]
-  } else {
-    0
-  }
+  # square <- if (x %in% trafos$var & any(trafos$fct == paste0(x, "^2"))) {
+  #   qdr_var <- trafos$Xc_var[which(trafos$var == x &
+  #                                          trafos$fct == paste0(x, "^2"))]
+  #   2 * as.numeric(MCMC[ , qdr_var])/scale_pars["scale", x]^2 * scale_pars["center", x]
+  # } else {
+  #   0
+  # }
 
-  new_vec <- vec / prod(scale_pars["scale", unlist(strsplit(x, ":"))]) - inter_sum - square
+  new_vec <- vec / prod(scale_pars["scale", unlist(strsplit(x, ":"))]) - inter_sum# - square
 
   new_vec
   # }