+++ NEWS.md
@@ -5,6 +5,7 @@
 * `simWide`: case with no observed bmi values removed
 * `traceplot()`, `densplot()`: `ncol` and `nrow` now work with `use_ggplot = TRUE`
 * functions with argument `subset` now return random effects covariance matrix correctly
+* `summary()` displayes output with rowname when only one node is returned
 
 ## Minor changes
 * `plot_all()` uses `xpd = TRUE` when printing text for character variables